learner/requirements.txt
flask-cors==3.0.9
flask-csv==1.2.0
flask-restful==0.3.9
gunicorn==20.1.0
pip-chill==1.0.1
pmdarima==1.8.0
pymannkendall==1.4.1
requests==2.25.1
rich==9.6.1
rope==0.18.0
seaborn==0.11.1
tqdm==4.53.0
